requirements.txt
python-telegram-bot==21.6
yt-dlp==2025.01.15
python-dotenv==1.0.1